Fix: Readmeの作成とディレクトリの細分化
