+++ .README_TEMPLATE.md
@@ -14,8 +14,7 @@
   - 1: 本番運用で必須
   - 2: 本番運用で推奨
   - 3: 開発環境でのみ必要
-  - 4: テスト環境でのみ必要
-  - 5: 無し
+  - 4: 不要
 
 理由: <本番運用での必要度の理由を記述>
 
@@ -25,7 +24,7 @@
   - 1: 開発/テスト運用で必須
   - 2: 開発/テスト運用で推奨
   - 3: 開発/テスト運用で任意
-  - 4: 無し
+  - 4: 不要
 
 理由: <開発/テスト運用での必要度の理由を記述>
