+++ internal/cli/README.md
@@ -1,6 +1,8 @@
 # CLI
 
-概要: このディレクトリは、アプリケーションのコマンドラインインターフェース (CLI) を提供します。データベースのマイグレーションやシード、サーバーの起動など、さまざまな操作を実行するためのコマンドを定義しています。
+概要: このディレクトリは、アプリケーションのコマンドラインインターフェース (CLI) を提供します。
+
+データベースのマイグレーションやシード、サーバーの起動など、さまざまな操作を実行するためのコマンドを定義しています。
 
 ## 役割
 
@@ -15,8 +17,7 @@
   - 1: 本番運用で必須
   - 2: 本番運用で推奨
   - 3: 開発環境でのみ必要
-  - 4: テスト環境でのみ必要
-  - 5: 無し
+  - 4: 無し
 
 理由: 本番環境では、データベースの管理やサーバーの起動に必要なため推奨されますが、必須ではありません。
