+++ internal/cli/migrate/README.md
@@ -1,0 +1,56 @@
+# Migrate
+
+概要: このディレクトリは、データベースのマイグレーション操作を管理するためのコマンドを提供します。
+
+`migrate-up` と `migrate-down` の2つの主要なコマンドを含みます。
+
+## 役割
+
+- データベースのスキーマを最新の状態に更新する (`migrate-up`)
+- 必要に応じてスキーマを以前の状態に戻す (`migrate-down`)
+
+## 必要度
+
+### 本番運用での必須度
+
+- 必須度: 1
+  - 理由: データベーススキーマの管理は本番環境でのアプリケーションの安定性に直結するため。
+
+### 開発/テスト運用での必須度
+
+- 必須度: 1
+  - 理由: 開発中やテスト中にスキーマを頻繁に変更する必要があるため。
+
+## MigrateUpの必要度
+
+### MigrateUpの本番運用での必須度
+
+- 必須度: 1
+  - 理由: 新しい機能や修正を反映するために、スキーマの更新が必要です。
+
+### MigrateUpの開発/テスト運用での必須度
+
+- 必須度: 1
+  - 理由: 開発中の変更を反映し、テスト環境での動作確認を行うため。
+
+## MigrateDownの必要度
+
+### MigrateDownの本番運用での必須度
+
+- 必須度: 3
+  - 理由: 通常、本番環境ではスキーマを戻す操作は推奨されません。ただし、緊急時には必要になる場合があります。
+
+### MigrateDownの開発/テスト運用での必須度
+
+- 必須度: 1
+  - 理由: 開発中やテスト中にスキーマを以前の状態に戻す必要があるため。
+
+### 無効化した場合の影響
+
+- スキーマの更新やロールバックができなくなり、アプリケーションの動作に支障をきたす可能性があります。
+
+## 注意点
+
+- `migrate-up` と `migrate-down` の操作は慎重に行ってください。
+- 本番環境での `migrate-down` は、データ損失のリスクがあるため、十分な検証とバックアップを行った上で実行してください。
+- マイグレーションファイルのパスは `database/migrations` に固定されています。
