+++ internal/cli/migrate/migratedown.go
@@ -1,4 +1,4 @@
-package cli
+package migrate
 
 import (
 	"errors"
