+++ internal/cli/migrate/migrateup.go
@@ -1,4 +1,4 @@
-package cli
+package migrate
 
 import (
 	"strconv"
