+++ internal/cli/seed/README.md
@@ -1,0 +1,32 @@
+# Seed
+
+概要: このディレクトリは、データベースに開発データを投入するためのコマンドを提供します。
+
+`db-seed` コマンドを使用して、シードファイルを実行します。
+
+## 役割
+
+- データベースに開発データを投入する。
+- テーブルが存在しない場合は、該当するシードファイルの実行をスキップします。
+
+## 必要度
+
+### 本番運用での必須度
+
+- 必須度: 4
+  - 理由: 本番環境では通常、開発データの投入は行われないため、シードデータの投入は不要です。ただし、初期データが必要な場合は手動で投入することがあります。
+
+### 開発/テスト運用での必須度
+
+- 必須度: 1
+  - 理由: 開発環境やテスト環境では、テストデータの投入が頻繁に必要となるため。
+
+### 無効化した場合の影響
+
+- 開発環境で初期データが投入されないため、アプリケーションの動作確認やテストが正しく行えなくなる可能性があります。
+
+## 注意点
+
+- シードファイルは `database/seed` ディレクトリに配置してください。
+- シードファイルが存在しないテーブルに対しては、エラーをスキップしますが、ログに警告が記録されます。
+- 本番環境での使用時には、投入するデータの内容を十分に確認してください。
