+++ internal/cli/serve/README.md
@@ -1,0 +1,32 @@
+# Serve
+
+概要: このディレクトリは、アプリケーションのサーバーを起動するためのコマンドを提供します。
+
+`serve` コマンドを使用して、アプリケーションのエントリーポイントを実行します。
+
+## 役割
+
+- アプリケーションのサーバーを起動し、リクエストを処理する。
+- 必要なミドルウェアやハンドラーを設定し、エンドポイントをバインドします。
+
+## 必要度
+
+### 本番運用での必須度
+
+- 必須度: 1
+  - 理由: サーバーの起動は本番環境でのアプリケーション運用に不可欠です。
+
+### 開発/テスト運用での必須度
+
+- 必須度: 1
+  - 理由: 開発中やテスト中にサーバーを起動して動作確認を行う必要があるため。
+
+### 無効化した場合の影響
+
+- サーバーが起動しないため、アプリケーションがリクエストを受け付けることができなくなります。
+
+## 注意点
+
+- サーバーのデフォルトポートは `:8080` に設定されています。必要に応じて変更してください。
+- データベース接続が必要な場合、設定ファイルで正しい接続情報を指定してください。
+- 本番環境では、適切なミドルウェア（例: ロギング、セキュリティ）を有効にすることを推奨します。
