+++ internal/config/README.md
@@ -19,8 +19,7 @@
   - 1: 本番運用で必須
   - 2: 本番運用で推奨
   - 3: 開発環境でのみ必要
-  - 4: テスト環境でのみ必要
-  - 5: 無し
+  - 4: 無し
 
 理由: 本番環境では、アプリケーションの設定が正確に読み込まれないと、サービス全体が正常に動作しなくなるため。
