+++ internal/controller/handler/error/response/README.md
@@ -55,8 +55,7 @@
   - 1: 本番運用で必須
   - 2: 本番運用で推奨
   - 3: 開発環境でのみ必要
-  - 4: テスト環境でのみ必要
-  - 5: 無し
+  - 4: 無し
 
 理由: 本番環境では、適切なエラーレスポンスを返すことで、クライアントにエラーの詳細を伝え、ユーザー体験を向上させるため必須です。
