+++ internal/controller/middleware/cors/README.md
@@ -15,8 +15,7 @@
   - 1: 本番運用で必須
   - 2: 本番運用で推奨
   - 3: 開発環境でのみ必要
-  - 4: テスト環境でのみ必要
-  - 5: 無し
+  - 4: 無し
 
 理由: 本番環境では、セキュリティを確保しつつ、特定のオリジンからのリクエストを許可するために必須です。
