+++ internal/controller/middleware/forcejson/README.md
@@ -15,8 +15,7 @@
   - 1: 本番運用で必須
   - 2: 本番運用で推奨
   - 3: 開発環境でのみ必要
-  - 4: テスト環境でのみ必要
-  - 5: 無し
+  - 4: 無し
 
 理由: 本番環境では、レスポンス形式の一貫性を保つことで、クライアント側の処理エラーを防ぐため推奨されます。
