+++ internal/controller/middleware/logging/README.md
@@ -15,8 +15,7 @@
   - 1: 本番運用で必須
   - 2: 本番運用で推奨
   - 3: 開発環境でのみ必要
-  - 4: テスト環境でのみ必要
-  - 5: 無し
+  - 4: 無し
 
 理由: 本番環境では、リクエストやレスポンスの詳細を記録することで、障害発生時のトラブルシューティングやモニタリングが容易になるため推奨されます。
