+++ internal/controller/server/README.md
@@ -14,8 +14,7 @@
   - 1: 本番運用で必須
   - 2: 本番運用で推奨
   - 3: 開発環境でのみ必要
-  - 4: テスト環境でのみ必要
-  - 5: 無し
+  - 4: 不要
 - 無効化した場合の影響: サーバーが正しく初期化されず、アプリケーション全体が動作しなくなる可能性があります
 
 ### binderの必要度
