+++ internal/infrastructure/rdb/README.md
@@ -21,8 +21,7 @@
   - 1: 本番運用で必須
   - 2: 本番運用で推奨
   - 3: 開発環境でのみ必要
-  - 4: テスト環境でのみ必要
-  - 5: 無し
+  - 4: 不要
 
 理由: 本番環境では、データベース接続やトランザクション管理が適切に行われないと、データの整合性が損なわれる可能性があるため必須です。
